# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/appointments": {
      "get": {
        "operationId": "AppointmentsController_getAppointments",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments"
          }
        },
        "summary": "List appointments",
        "tags": [
          "Appointments"
        ]
      },
      "post": {
        "operationId": "AppointmentsController_createAppointment",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully"
          }
        },
        "summary": "Create an appointment",
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/user/{userId}": {
      "get": {
        "operationId": "AppointmentsController_getAppointmentsByUserId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user appointments"
          }
        },
        "summary": "Get appointments for a user",
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/completed/{userId}": {
      "get": {
        "operationId": "AppointmentsController_getCompletedAppointmentsByUserId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of completed user appointments"
          }
        },
        "summary": "Get completed appointments for a user",
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/doctor/{doctorId}": {
      "get": {
        "operationId": "AppointmentsController_getAppointmentsByDoctorId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of doctor appointments"
          }
        },
        "summary": "Get appointments for a doctor",
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/{id}": {
      "put": {
        "operationId": "AppointmentsController_updateAppointment",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated successfully"
          }
        },
        "summary": "Update an appointment",
        "tags": [
          "Appointments"
        ]
      },
      "delete": {
        "operationId": "AppointmentsController_cancelAppointment",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment canceled successfully"
          }
        },
        "summary": "Cancel an appointment",
        "tags": [
          "Appointments"
        ]
      }
    },
...
  },
  "info": {
    "title": "MEDBITS Healthcare API",
    "description": "REST API for the MEDBITS Healthcare Management System.\n\n**Authentication**: Pass the user role in the `role` request header.\n\nValid roles: `patient` | `doctor` | `frontdesk` | `admin`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC"
      }
    },
    "schemas": {
...
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "PAT001"
          },
          "doctorId": {
            "type": "string",
            "example": "DOC001"
          },
          "date": {
            "type": "string",
            "example": "2023-12-01"
          },
          "slot": {
            "type": "string",
            "example": "10:00 AM"
          }
        },
        "required": [
          "userId",
          "doctorId",
          "date",
          "slot"
        ]
      },
      "UpdateAppointmentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2023-12-02"
          },
          "slot": {
            "type": "string",
            "example": "11:00 AM"
          }
        }
      },
...
    }
  }
}